+++ packages/vite-plugin-super-template/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.5
+
+- Upgrade to `@visma/react-intl-bundled-messages@^1.3.0`
+
 # 0.1.4
 
 - Add `vite-plugin-dynamic-base`